requirements.txt
ujson==4.2.0
django==4.0.1
hypercorn==0.13.2
psycopg2==2.9.3
djangorestframework==3.13.1
boto3==1.24.46
django-model-utils==4.2.0
drf-nested-routers==0.93.4
